requirements.txt
folium==0.19.4
haversine==2.9.0
numpy==2.2.3
pandas==2.2.3
Pillow==11.1.0
plotly==6.0.0
streamlit==1.42.2
streamlit_folium==0.24.0
